Revert "Join Fixture extensions functions in the same file to improve import usage"
This reverts commit fd1af581
