+++ kotlin-fixture/src/main/java/com/marcellogalhardo/fixture/FixtureRegistry.kt
@@ -1,0 +1,12 @@
+package com.marcellogalhardo.fixture
+
+import kotlin.reflect.KClass
+
+interface FixtureRegistry {
+
+    fun <T : Any> register(classRef: KClass<T>, providerFunction: ProviderFunction<T>)
+}
+
+inline fun <reified T : Any> FixtureRegistry.register(noinline providerFunction: ProviderFunction<T>) {
+    register(T::class, providerFunction)
+}
